requirement.txt
boto3==1.42.82
botocore==1.42.82
chromadb==1.5.5
huggingface_hub==1.9.0
langchain==1.2.12
langchain-aws==1.4.3
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.25
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.1
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.16
uvicorn==0.41.0
streamlit==1.56.0
pandas==3.0.2
python-dotenv